services/rag-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
qdrant-client==1.11.1
pydantic==2.9.2
httpx==0.27.2
python-docx==1.1.2
pymupdf==1.24.10
markdown==3.7